+++ Framework/BGThemeManager.m
@@ -107,7 +107,7 @@ static BGThemeManager *keyedManager = nil;
     return self;
 }
 
-- (unsigned)retainCount; {
+- (NSUInteger)retainCount; {
     return UINT_MAX;  //denotes an object that cannot be released
 }
